perf(www): landing page is an RSC, template picker is the only island
PERF-102. 'use client' sat at module scope on the 652-line landing page,
so the feature grid, step copy, and four hand-written syntax-highlighted
code blocks (153 span-producing helper calls) were compiled into the
route chunk and hydrated even though the only client state is the
5-button template picker's useState.

Extract the picker into www/components/template-picker.tsx; the page
keeps its static markup server-only.
